architecture in prog

+++ Scripts/Datas/TowerData.cs
@@ -8,7 +8,7 @@ namespace Datas
     [Serializable]
     public class TowerData
     {
-        public int Height;
+        public float Height;
         public TeamData TeamData;
     }
 }
